+++ tests/integration/fetch_run_details_tests.rs
@@ -166,7 +166,7 @@ fn returns_reportable_entities_if_no_reportable_entities_flag() {
         .stdout(predicate::str::contains("reportable@test.com").not());
 }
 
-fn build_run(conn: &Connection, index: u8) -> u32 {
+fn build_run(conn: &Connection, index: u8) -> i64 {
     let persisted_source = persist_message_source(conn, message_source(index));
 
     let output_data = build_output_data(persisted_source);
